Module yew::agent

This module contains types to support multi-threading and state management.

Structs

AgentLink
Link to agent's scope for creating callbacks.

Context
Create a single instance in the current thread.

Dispatcher
A newtype around a bridge to indicate that it is distinct from a normal bridge

HandlerId
Id of responses handler.

Job
Create an instance in the current thread.

Private
Create a new instance for every bridge.

Public
Create a single instance in a tab.

Traits

Agent
Declares the behavior of the agent.

Bridge
Bridge to a specific kind of worker.

Bridged
This trait allows registering or getting the address of a worker.

Dispatched
This trait allows the creation of a dispatcher to an existing agent that will not send replies when messages are sent.

Threaded
Implements rules to register a worker in a separate thread.