Trait yew::agent::Bridged

pub trait Bridged: Agent + Sized + 'static {
    fn bridge(callback: Callback<Self::Output>) -> Box<dyn Bridge<Self>>;
}

This trait allows registering or getting the address of a worker.

Required methods

fn bridge(callback: Callback<Self::Output>) -> Box<dyn Bridge<Self>>

Creates a messaging bridge between a worker and the component.

Loading content...