Function yew::set_custom_panic_hook

pub fn set_custom_panic_hook(
    hook: Box<dyn Fn(&PanicInfo<'_>) + Sync + Send + 'static>
)

Set a custom panic hook. Unless a panic hook is set through this function, Yew will overwrite any existing panic hook when an application is rendered with Renderer.