Module yew::format

Utility module to convert data to types and back by specific formats like: JSON, BSON, TOML, YAML, XML.

All types here are lazy and it's necessary to use Into and From traits to get (convert) the data.

Re-exports

pub use self::nothing::Nothing;

pub use self::json::Json;

Modules

json
Contains an implementation of JSON serialization format.

macros
Contains macro for wrapping serde format.

nothing
Contains an implementation of empty serialization format (Nothing).

Type Definitions

Binary
A representation of a value which can be stored and restored as a binary.

Text
A representation of a value which can be stored and restored as a text.