Trait yew::html::ImplicitClone

pub trait ImplicitClone: Clone { }

Marker trait for types that can be cloned implicitly.

Behaves exactly like Copy but calls the Clone implementation instead and must be implemented in the host library.