Trait yew::html::IntoOptPropValue

pub trait IntoOptPropValue<T> {
    fn into_opt_prop_value(self) -> Option<T>;
}

A trait similar to Into<Option<T>> which allows conversion to an optional value of a Properties struct.

Required methods

fn into_opt_prop_value(self) -> Option<T>

Convert self to an optional value of a Properties struct.