Trait yew::html::ToHtml

pub trait ToHtml {
    // Required method
    fn to_html(&self) -> Html;

    // Provided method
    fn into_html(self) -> Html
       where Self: Sized { ... }
}

A trait implemented for types be rendered as a part of a Html.

Types that implements this trait can define a virtual dom layout that itself should be rendered into via html! and can be referenced / consumed as {value} in an html! macro invocation.

Required Methods

fn to_html(&self) -> Html

Converts this type to a Html.

Provided Methods

fn into_html(self) -> Html
where Self: Sized,

Converts this type into a Html.