Macro yew::text_format_is_an_error

macro_rules! text_format_is_an_error {
    ($type:ident) => { ... };
}

This macro is used for a format that can be encoded as Binary but can't be encoded as Text. It is used in conjunction with a type definition for a tuple struct with one (publically accessible) element of a generic type. This macro should be paired with the binary_format macro that defines the binary-only format.

Example

  use rmp_serde;
  use yew::{binary_format, text_format_is_an_error};

  pub struct MsgPack<T>(pub T);

  binary_format!(MsgPack based on rmp_serde);
  text_format_is_an_error!(MsgPack);