Function yew::platform::pinned::mpsc::unbounded

pub fn unbounded<T>() -> (UnboundedSender<T>, UnboundedReceiver<T>)

Creates an unbounded channel.

The send method on Senders created by this function will always succeed and return immediately as long as the channel is open.

Note

This channel has an infinite buffer and can run out of memory if the channel is not actively drained.