Struct yew::prelude::Classes

pub struct Classes { /* private fields */ }

A set of classes.

The preferred way of creating this is using the classes! macro.

Implementations

impl Classes

pub fn new() -> Self

Creates an empty set of classes. (Does not allocate.)

pub fn with_capacity(n: usize) -> Self

Creates an empty set of classes with capacity for n elements. (Does not allocate if n is zero.)

pub fn push<T: Into<Self>>(&mut self, class: T)

Adds a class to a set.

If the provided class has already been added, this method will ignore it.

pub unsafe fn unchecked_push<T: Into<Cow<'static, str>>>(&mut self, class: T)

Adds a class to a set.

If the provided class has already been added, this method will ignore it.

This method wonโt check if there are multiple classes in the input string.

Safety

This function will not split the string into multiple classes. Please do not use it unless you are absolutely certain that the string does not contain any whitespace and it is not empty. Using push() is preferred.

pub fn contains<T: AsRef<str>>(&self, class: T) -> bool

Check the set contains a class.

pub fn is_empty(&self) -> bool

Check the set is empty.