Module yew::services

This module is a container of servies to interact with the external resources.

It carries a similar role as subscriptions in Elm, but can be used directly from the update method.

Modules

console
This module contains a service implementation to use browser's console.

dialog
This module contains the implementation of a service to show alerts and confirm dialogs in a browser.

fetch
Service to send HTTP-request to a server.

interval
This module contains the implementation of a service for periodic sending messages to a loop.

keyboard
Service to register key press event listeners on elements.

reader
Service to load files using FileReader.

render
This module contains Yew's implementation of a service which can be used to request frame rendering

resize
This module contains Yew's implementation of a service which listens for browser window resize events.

storage
This module contains Yew's implementation of a service to use local and session storage of a browser.

timeout
This module contains Yew's implementation of a service to send messages when a timeout has elapsed.

websocket
A service to connect to a server through the WebSocket Protocol.

Structs

ConsoleService
A service to use methods of a Console.

DialogService
A dialog service.

FetchService
A service to fetch resources.

IntervalService
A service to send messages on every elapsed interval.

ReaderService
A reader service attached to a user context.

RenderService
A service to request animation frames.

ResizeService
A service which fires events when the browser window is resized.

StorageService
A storage service attached to a context.

TimeoutService
An service to set a timeout.

WebSocketService
A WebSocket service attached to a user context.

Traits

Task
A task is an ongoing process which is part of a Yew application.