Struct yew::AppHandle

pub struct AppHandle<COMP: BaseComponent> { /* private fields */ }

An instance of an application.

Implementations

impl<COMP> AppHandle<COMP>where
COMP: BaseComponent,

pub fn update(&mut self, new_props: COMP::Properties)

Update the properties of the app’s root component.

This can be an alternative to sending and handling messages. The existing component will be reused and have its properties updates. This will presumably trigger a re-render, refer to the changed lifecycle for details.

pub fn destroy(self)

Schedule the app for destruction