Module yew::suspense

This module provides suspense support.

Structs

Suspense

Suspend rendering and show a fallback UI until the underlying task completes.

SuspenseProps

Properties for Suspense.

Suspension

A Suspension.

SuspensionHandle

A Suspension Handle.

UseFutureHandle

This hook is used to await a future in a suspending context.

Functions

use_future

Use the result of an async computation, suspending while waiting.

use_future_with

Use the result of an async computation with dependencies, suspending while waiting.

Type Aliases

SuspensionResult

A Suspension Result.