Module yew::utils

This module contains useful utilities to get information about the current document.

Structs

NodeSeq

A special type necessary for flattening components returned from nested html macros.

Functions

into_node_iter

Map IntoIterator<Item = Into<T>> to Iterator<Item = T>

print_node

Print the web_sys::Node’s contents as a string for debugging purposes