Trait yew::virtual_dom::Listener

pub trait Listener {
    fn kind(&self) -> ListenerKind;
    fn handle(&self, event: Event);
    fn passive(&self) -> bool;
}

The Listener trait is an universal implementation of an event listener which is used to bind Rust-listener to JS-listener (DOM).

Required Methods

fn kind(&self) -> ListenerKind

Returns the name of the event

fn handle(&self, event: Event)

Handles an event firing

fn passive(&self) -> bool

Makes the event listener passive. See addEventListener.