Trait yew::virtual_dom::vcomp::Transformer

pub trait Transformer<COMP: Component, FROM, TO> {
    fn transform(&mut self, from: FROM) -> TO;
}

Converts property and attach lazy components to it. This type holds context and components types to store an activatior which will be used later buring rendering state to attach component sender.

Required methods

fn transform(&mut self, from: FROM) -> TO

Transforms one type to another.

Loading content...