Struct yew::virtual_dom::vtag::VTag

pub struct VTag<COMP: Component> {
    pub reference: Option<Element>,
    pub listeners: Vec<Box<dyn Listener<COMP>>>,
    pub attributes: HashMap<String, String>,
    pub childs: Vec<VNode<COMP>>,
    pub classes: HashSet<String>,
    pub value: Option<String>,
    pub kind: Option<String>,
    pub checked: bool,
    // some fields omitted
}

A type for a virtual Element representation.

Fields
reference: Option<Element>

A reference to the Element.

listeners: Vec<Box<dyn Listener<COMP>>>

List of attached listeners.

attributes: HashMap<String, String>

List of attributes.

childs: Vec<VNode<COMP>>

The list of children nodes. Which also could have own children.

classes: HashSet<String>

List of attached classes.

value: Option<String>

Contains a value of an InputElement.

kind: Option<String>

Contains kind value of an InputElement.

checked: bool

Represents checked attribute of input. It exists to override standard behavior of checked attribute, because in original HTML it sets defaultChecked value of InputElement, but for reactive frameworks it's more useful to control checked value of an InputElement.

Methods

impl<COMP: Component> VTag<COMP>

pub fn new<S: Into<Cow<'static, str>>>(tag: S) -> Self

Creates a new VTag instance with tag name (cannot be changed later in DOM).

pub fn tag(&self) -> &str

Returns tag of an Element. In HTML tags are always uppercase.

pub fn add_child(&mut self, child: VNode<COMP>)

Add VNode child.

pub fn add_class(&mut self, class: &str)

Adds asingle class to this virtual node. Actually it will set by Element.classList.add call later.

pub fn set_classes(&mut self, classes: &str)

Add classes to this virtual node. Actually it will set by Element.classList.add call later.

pub fn set_value<T: ToString>(&mut self, value: &T)

Sets value for an InputElement.

pub fn set_kind<T: ToString>(&mut self, value: &T)

Sets kind property of an InputElement. Same as set type attribute.

pub fn set_checked(&mut self, value: bool)

Sets checked property of an InputElement. (Not a value of node's attribute).

pub fn add_attribute<T: ToString>(&mut self, name: &str, value: &T)

Adds attribute to a virtual node. Not every attribute works when it set as attribute. We use workarounds for: class, type/kind, value and checked.

pub fn add_listener(&mut self, listener: Box<dyn Listener<COMP>>)

Adds new listener to the node. It's boxed because we want to keep it in a single list. Lates Listener::attach called to attach actual listener to a DOM node.