Module yew::agent

This module contains types to support multi-threading in Yew.

Structs

AgentLink
Link to agent's scope for creating callbacks.

AgentScope
This sctruct holds a reference to a component and to a global scheduler.

Context
Create a single instance in the current thread.

Global
Create a single instance in a browser.

HandlerId
Id of responses handler.

Job
Create an instance in the current thread.

Private
Create a new instance for every bridge.

PrivateBridge
A connection manager for components interaction with workers.

Public
Create a single instance in a tab.

PublicBridge
A connection manager for components interaction with workers.

Traits

Agent
Declares the behavior of the agent.

Bridge
Bridge to a specific kind of worker.

Bridged
This traits allow to get addres or register worker.

Threaded
Implements rules to register a worker in a separate thread.

Transferable
Represents a message which you could send to an agent.