Module yew::html

The main module which contents aliases to necessary items to create a template and implement update and view functions. Also this module contains declaration of Component trait which used to create own UI-components.

Modules

onblur
An abstract implementation of a listener.

onchange
An abstract implementation of a listener.

onclick
An abstract implementation of a listener.

oncontextmenu
An abstract implementation of a listener.

ondoubleclick
An abstract implementation of a listener.

ondrag
An abstract implementation of a listener.

ondragend
An abstract implementation of a listener.

ondragenter
An abstract implementation of a listener.

ondragexit
An abstract implementation of a listener.

ondragleave
An abstract implementation of a listener.

ondragover
An abstract implementation of a listener.

ondragstart
An abstract implementation of a listener.

ondrop
An abstract implementation of a listener.

onfocus
An abstract implementation of a listener.

ongotpointercapture
An abstract implementation of a listener.

oninput
An abstract implementation of a listener.

onkeydown
An abstract implementation of a listener.

onkeypress
An abstract implementation of a listener.

onkeyup
An abstract implementation of a listener.

onlostpointercapture
An abstract implementation of a listener.

onmousedown
An abstract implementation of a listener.

onmouseenter
An abstract implementation of a listener.

onmouseleave
An abstract implementation of a listener.

onmousemove
An abstract implementation of a listener.

onmouseout
An abstract implementation of a listener.

onmouseover
An abstract implementation of a listener.

onmouseup
An abstract implementation of a listener.

onmousewheel
An abstract implementation of a listener.

onpointercancel
An abstract implementation of a listener.

onpointerdown
An abstract implementation of a listener.

onpointerenter
An abstract implementation of a listener.

onpointerleave
An abstract implementation of a listener.

onpointermove
An abstract implementation of a listener.

onpointerout
An abstract implementation of a listener.

onpointerover
An abstract implementation of a listener.

onpointerup
An abstract implementation of a listener.

onscroll
An abstract implementation of a listener.

onsubmit
An abstract implementation of a listener.

Structs

ComponentLink
Link to component's scope for creating callbacks.

Href
A bridging type for checking href attribute value.

InputData
A type representing data from oninput event.

Scope
A context which contains a bridge to send a messages to a loop. Mostly services uses it.

Enums

ChangeData
A type representing change of value(s) of an element after committed by user (onchange event).

Traits

Component
An interface of a UI-component. Uses self as a model.

Renderable
Should be rendered relative to context and component environment.

Type Definitions

Html
A type which expected as a result of view function implementation.

NodeCell
Holder for the element.

ShouldRender
This type indicates that component should be rendered again.