Module yew::html

The main html module which defines components, listeners, and class helpers.

Structs

AnyScope

Untyped scope used for accessing parent scope

ChildrenRenderer

A type used for rendering children html.

Classes

A set of classes.

InputData

A type representing data from oninput event.

NodeRef

Wrapped Node reference for later use in Component lifecycle methods.

Scope

A context which allows sending messages to a component.

Enums

ChangeData

A type representing change of value(s) of an element after committed by user (onchange event).

Traits

Component

Components are the basic building blocks of the UI in a Yew app. Each Component chooses how to display itself using received props and self-managed state. Components can be dynamic and interactive by declaring messages that are triggered and handled asynchronously. This async update mechanism is inspired by Elm and the actor model used in the Actix framework.

ImplicitClone

Marker trait for types that the [html!] macro may clone implicitly.

IntoOptPropValue

A trait similar to Into<Option<T>> which allows conversion to an optional value of a Properties struct.

IntoPropValue

A trait similar to Into<T> which allows conversion to a value of a Properties struct.

Properties

Trait for building properties for a component

SendAsMessage

Defines a message type that can be sent to a component. Used for the return value of closure given to Scope::batch_callback.

Type Definitions

Children

A type used for accepting children elements in Component::Properties.

ChildrenWithProps

A type used for accepting children elements in Component::Properties and accessing their props.

ComponentLink

Link to component’s scope for creating callbacks.

Html

A type which expected as a result of view function implementation.

ShouldRender

This type indicates that component should be rendered again.

Derive Macros

Properties