Trait yew::html::Component

pub trait Component: Sized + 'static {
    type Message: 'static;
    type Properties: Clone + PartialEq + Default;
    fn create(props: Self::Properties, link: ComponentLink<Self>) -> Self;

    fn update(&mut self, msg: Self::Message) -> ShouldRender;

    fn change(&mut self, _: Self::Properties) -> ShouldRender { ... }

    fn destroy(&mut self) { ... }
}

An interface of a UI-component. Uses self as a model.

Associated Types

type Message: 'static

Control message type which update loop get.

type Properties: Clone + PartialEq + Default

Properties type of component implementation. It sould be serializable because it's sent to dynamicaly created component (layed under VComp) and must be restored for a component with unknown type.

Loading content...
Required methods

fn create(props: Self::Properties, link: ComponentLink<Self>) -> Self

Initialization routine which could use a context.

fn update(&mut self, msg: Self::Message) -> ShouldRender

Called everytime when a messages of Msg type received. It also takes a reference to a context.

Loading content...
Provided methods

fn change(&mut self, _: Self::Properties) -> ShouldRender

This method called when properties changes, and once when component created.

fn destroy(&mut self)

Called for finalization on the final point of the component's lifetime.

Loading content...