Trait yew::html::IntoPropValue

pub trait IntoPropValue<T> {
    fn into_prop_value(self) -> T;
}

A trait similar to Into<T> which allows conversion to a value of a Properties struct.

Required Methods

fn into_prop_value(self) -> T

Convert self to a value of a Properties struct.