Trait yew::html::SendAsMessage

pub trait SendAsMessage<COMP: Component> {
    fn send(self, scope: &Scope<COMP>);
}

Defines a message type that can be sent to a component. Used for the return value of closure given to Scope::batch_callback.

Required methods

fn send(self, scope: &Scope<COMP>)

Sends the message to the given component’s scope. See Scope::batch_callback.