Module yew::prelude

The Yew Prelude

The purpose of this module is to alleviate imports of many common types:

use yew::prelude::*;

Re-exports

pub use crate::agent::Bridge;

pub use crate::agent::Bridged;

pub use crate::app::App;

pub use crate::callback::Callback;

pub use crate::events::*;

pub use crate::html::Children;

pub use crate::html::ChildrenWithProps;

pub use crate::html::Component;

pub use crate::html::ComponentLink;

pub use crate::html::Html;

pub use crate::html::NodeRef;

pub use crate::html::Properties;

pub use crate::html::ShouldRender;

pub use crate::macros::html;

Modules

worker
Prelude module for creating worker.

Macros

classes

html

html_nested

Structs

Classes
A set of classes.

Traits

Dispatched
This trait allows the creation of a dispatcher to an existing agent that will not send replies when messages are sent.

Threaded
Implements rules to register a worker in a separate thread.

Derive Macros

Properties