Module yew::prelude

The Yew Prelude

The purpose of this module is to alleviate imports of many common types:

use yew::prelude::*;

Re-exports

pub use crate::app::App;

pub use crate::callback::Callback;

pub use crate::events::*;

pub use crate::html::Html;

pub use crate::html::NodeRef;

pub use crate::macros::html;

Macros

classes

html

html_nested

Structs

Classes

A set of classes.

Traits

Component

Components are the basic building blocks of the UI in a Yew app. Each Component chooses how to display itself using received props and self-managed state. Components can be dynamic and interactive by declaring messages that are triggered and handled asynchronously. This async update mechanism is inspired by Elm and the actor model used in the Actix framework.

Properties

Trait for building properties for a component

Type Definitions

Children

A type used for accepting children elements in Component::Properties.

ChildrenWithProps

A type used for accepting children elements in Component::Properties and accessing their props.

ComponentLink

Link to component’s scope for creating callbacks.

ShouldRender

This type indicates that component should be rendered again.

Derive Macros

Properties