Module yew::services

This module is a container of servies to interact with the external resources.

It carries a similar role as subscriptions in Elm, but can be used directly from the update method.

Re-exports

pub use self::console::ConsoleService;

pub use self::dialog::DialogService;

pub use self::fetch::FetchService;

pub use self::interval::IntervalService;

pub use self::reader::ReaderService;

pub use self::render::RenderService;

pub use self::storage::StorageService;

pub use self::timeout::TimeoutService;

pub use self::websocket::WebSocketService;

Modules

console
This module contains a service implementation to use browser's console.

dialog
This module contains the implementation of a service to show alerts and confirm dialogs in a browser.

fetch
Service to send HTTP-request to a server.

interval
This module contains the implementation of a service for periodic sending messages to a loop.

reader
Service to load files using FileReader.

render
This module contains the implementation of a service to request frame rendering

storage
This module contains the implementation of a service to use local and session storage of a browser.

timeout
This module contains the implementation of a service to send a messages when timeout elapsed.

websocket
Service to connect to a servers by WebSocket Protocol.

Traits

Task
An universal task of a service. It have to be canceled when dropped.