Module yew::virtual_dom

This module contains Yew's implementation of a reactive virtual DOM.

Structs

Key
Represents the (optional) key of Yew's virtual nodes.

PositionalAttr
Key-value tuple which makes up an item of the Attributes::Vec variant.

VChild
A virtual child component.

VComp
A virtual component.

VList
This struct represents a fragment of the Virtual DOM tree.

VTag
A type for a virtual Element representation.

VText
A type for a virtual TextNode representation.

Enums

Attributes
A collection of attributes for an element

VNode
Bind virtual element to a DOM reference.

Traits

Listener
The Listener trait is an universal implementation of an event listener which is used to bind Rust-listener to JS-listener (DOM).

Transformer
Transform properties to the expected type.