Module yew::virtual_dom

This module contains the implementation of reactive virtual dom concept.

Re-exports

pub use self::vcomp::VComp;

pub use self::vlist::VList;

pub use self::vnode::VNode;

pub use self::vtag::VTag;

pub use self::vtext::VText;

Modules

vcomp
This module contains the implementation of a virtual component VComp.

vlist
This module contains fragments implementation.

vnode
This module contains the implementation of abstract virtual node.

vtag
This module contains the implementation of a virtual element node VTag.

vtext
This module contains the implementation of a virtual text node VText.

Traits

Listener
Listener trait is an universal implementation of an event listener which helps to bind Rust-listener to JS-listener (DOM).

VDiff
This trait provides features to update a tree by other tree comparsion.