Module yew::virtual_dom

This module contains Yew’s implementation of a reactive virtual DOM.

Structs

Key

Represents the (optional) key of Yew’s virtual nodes.

VChild

A virtual child component.

VComp

A virtual component.

VList

This struct represents a fragment of the Virtual DOM tree.

VPortal

VRaw

A raw HTML string to be used in VDOM.

VSuspense

This struct represents a suspendable DOM fragment.

VTag

A type for a virtual Element representation.

VText

A type for a virtual TextNode representation.

Enums

ApplyAttributeAs

Defines if the Attributes is set as element’s attribute or property

Attributes

A collection of attributes for an element

ListenerKind

Supported kinds of DOM event listeners

Listeners

A list of event listeners

VNode

Bind virtual element to a DOM reference.

Traits

Listener

The Listener trait is an universal implementation of an event listener which is used to bind Rust-listener to JS-listener (DOM).

Type Definitions

AttrValue

Attribute value