Trait yew::virtual_dom::Listener

pub trait Listener<COMP: Component> {
    fn kind(&self) -> &'static str;

    fn attach(
        &mut self, 
        element: &Element, 
        scope: Scope<COMP>
    ) -> EventListenerHandle;
}

Listener trait is an universal implementation of an event listener which helps to bind Rust-listener to JS-listener (DOM).

Required methods

fn kind(&self) -> &'static str

Returns standard name of DOM's event.

fn attach(
&mut self,
element: &Element,
scope: Scope<COMP>
) -> EventListenerHandle

Attaches listener to the element and uses scope instance to send prepaired event back to the yew main loop.

Loading content...